+++ tests/test_core.py
@@ -485,6 +485,51 @@ def test_config_invalid_meta():
         config_parser(config)
 
 
+def test_config_environment_variables():
+    """Test the case where environment variables are passed to a stage."""
+    config = {
+        'build-magic': [
+            {
+                'stage': {
+                    'commands': [
+                        {'execute': "echo '$HELLO' '$WORLD'"}
+                    ],
+                    'environment variables': {
+                        'HELLO': 'hello',
+                        'WORLD': 'world',
+                    }
+                }
+            }
+        ]
+    }
+    ref = [
+        {
+            'name': '',
+            'runner_type': 'local',
+            'environment': '',
+            'continue': False,
+            'wd': '.',
+            'copy': '',
+            'artifacts': [],
+            'action': 'default',
+            'commands': [
+                "echo '$HELLO' '$WORLD'",
+            ],
+            'directives': [
+                'execute',
+            ],
+            'dotenv': '',
+            'parameters': [],
+            'environment variables': {
+                'HELLO': 'hello',
+                'WORLD': 'world',
+            }
+        }
+    ]
+    stages = config_parser(config)
+    assert stages == ref
+
+
 def test_config_parser_validation_fail():
     """Test the case where config file schema validation fails."""
     config = {