+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added support for job and stage descriptions.
 * An environment can now be specified for the Local command runner and the stage will be skipped if the environment doesn't match the current OS or distribution.
 * Build-magic creates and uses a custom Vagrantfile when using the Vagrant command runner and setting bind directory or envvars.
+* Bugfixes
 
 ## Version 0.3.3 (Newest)